move data clases to a separate file.
